==> Notebooks/1-Get_Sentiment_Description.ipynb <==
import pandas as pd
import numpy as np
import re
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer
from nltk.tokenize import RegexpTokenizer
from nltk.corpus import stopwords

#TextBlob
from textblob import TextBlob

#SpaCy
import spacy
from spacy import displacy

from langdetect import detect


# Reading the csv file and storing it in a dataframe called df.
df = pd.read_csv("../data/netflix_titles.csv")
df.head()


def tokenize (string):
    # Tokenizing the string.
    try:
        tokenizer = RegexpTokenizer(r"\w+")
        tokens = tokenizer.tokenize(string)
        return tokens
    # Returning a NaN value if the tokenizer fails to tokenize the string.
    except: 
        return np.nan

    
def stop_words(list_):
    #Creating a list of stop words in english.
    stop_words = stopwords.words("english")

    new_lista = []
    
    # Iterating through the list of words and if the word is not in the stop words list, it will append it to the new list.
    for string in list_: 
        if string.lower() not in stop_words:
            new_lista.append(string)
    return " ".join(new_lista)

def sentiment(string):
    #Creating an object of the class SentimentIntensityAnalyzer.
    sia = SentimentIntensityAnalyzer()
   
    # Creating a dictionary with the polarity of the string.
    polarity = sia.polarity_scores(string)
    
    # Returning the compound value of the polarity dictionary.
    pol = polarity["compound"]
    return pol

# apply the function to tokenize the data
df["descr_tok"] = df["description"].apply(tokenize)

df.head(2)

# apply the function to get the description without the stopwords
df["descr_tok"] = df["descr_tok"].apply(stop_words)

# Downloading the vader lexicon from the nltk library.
nltk.download("vader_lexicon")

# apply the function to get the sentiment analysis values
df["sentiment"] = df["descr_tok"].apply(sentiment)

df.head()

# Saving the dataframe to a csv file.
df.to_csv("../data/netflix_sentiment.csv")
