==> Notebooks/4-Netflix_Originals.ipynb <==
import pandas as pd
import numpy as np
import re
import pickle

# Adding the parent directory to the path so that the `src` directory can be imported.
import sys
sys.path.append('../')

# Importing the `support` module from the `src` directory.
import src.support as sp

# Reading the csv file and storing it in a dataframe called df.
originals = pd.read_csv("../data/netflix_originals.csv", index_col = 0)
originals.head()

# Reading the csv file and storing it in a dataframe called df.
df= pd.read_csv("../data/netflix_titles.csv")
df.head(1)

# merging two datasets
originals_df = df.merge(originals, left_on= "title", right_on="Title", how="inner")

# Dropping the columns that are not needed for the analysis.
originals_df.drop(['release_year', 'rating', 'duration', 'listed_in', 'description','Title', "date_added"], axis = 1, inplace=True)
originals_df.head(2)

# Dropping all rows that have a NaN value in any column.
originals_df2 = originals_df.dropna()

# Cleaning

## Country

# get the main production country
originals_df["country"] = originals_df["country"].apply(lambda x: x.split(",")[0] if type(x) == str else x)
originals_df.head(1)

## Language

# get the main language

originals_df["Language"] = originals_df["Language"].apply(lambda x: x.split("/")[0] if type(x) == str else x)
originals_df.head(2)

## Cast

# extracting all the actors that participate in the films, and separate each actor in differents rows
originals_df["cast"] = originals_df["cast"].apply(lambda x: x.split(",") if type(x) == str else x)
all_cast = originals_df.explode("cast")
all_cast["cast"] = all_cast["cast"].str.strip()
originals_df.head()

## Director

# getting the most populats directors
top10_directors = all_cast.groupby("director").size().reset_index().sort_values(by = 0, ascending = False)["director"].iloc[:10].to_list()
top10_directors

# The above code is opening the pickle file and writing the top10_directors dataframe to the pickle file.
with open ('../data/pickle/top10_directors.pickle', "wb") as dir:
    pickle.dump(top10_directors, dir)

## Genre

# extracting all the genres that participate in the films, and separate each actor in differents rows
originals_df["Genre"] = originals_df["Genre"].apply(lambda x: x.split("/")[0].split("-")[0].replace("film", "").replace("thriller", "").strip() if type(x) == str else x)
originals_df.head(2)

def clean_genre(x):
    
    drama = re.findall("drama", x)
    romantic = re.findall("Roman", x)
    dance = re.findall("[D|d]ance|[M|m]us",x)
    comedy = re.findall("[C|c]ome", x)
    nulos = re.findall("^$|One", x)

    if drama:
        return ""
    elif dance:
        return "Musical"
    elif romantic:
        return "Romantic"
    elif comedy:
        return "Comedy"
    
    elif nulos:
        return np.nan
    else:
        return x

    

# aplying the clean_genre function
originals_df["Genre"] = originals_df["Genre"].apply(clean_genre)

originals_df.groupby(["director"])["IMDB Score"].size().reset_index().sort_values(by = "IMDB Score", ascending = False)

originals_df.to_csv("../data/netflix_originals_clean.csv")
