==> Notebooks/5-Srapp_Actors_Directors.ipynb <==
import pandas as pd
import numpy as np
import re
import pickle

# Adding the parent directory to the path so that the `src` directory can be imported.
import sys
sys.path.append("../")

# Importing the `support` module from the `src` directory.
import src.support as sp

# Ignoring the warnings.
import warnings
warnings.filterwarnings('ignore')

# open the top10 actors pickle file
with open('../data/pickle/top10_actors.pickle', 'rb') as act:
    actors = pickle.load(act)

# scrapping personal information for each actor from IMDB
actors_df = sp.scrap_actors(actors)

actors_df.head()

actors_df["known_for"] = actors_df["known_for"].str.split("\n")

actors_df = actors_df.explode("known_for")

actors["description"] = actors["description"].str.replace(",", "")

actors_df.to_csv("../data/top10_actors.csv", index = False, sep = ";")


# Directors

# open the top10 directors pickle file
with open('../data/pickle/top10_directors.pickle', 'rb') as dir:
    directors = pickle.load(dir)

# scrapping personal information for each actor from IMDB
directors_df  = sp.scrap_directors(directors)

directors_df["description"] = directors_df["description"].str.replace(",", "").replace(";", "")

directors_df["description"][2]

directors_df.to_csv("../data/top10_directors.csv", index=False, sep= ";")
