==> Notebooks/6-Insert_Data_SQL.ipynb <==
import pandas as pd
import sqlalchemy as alch

# Importing the getpass function from the getpass module.
from getpass import getpass 

# Reading the csv file and storing it in a dataframe called df.
actors = pd.read_csv("../data/top10_actors.csv", sep = ";")
directors = pd.read_csv("../data/top10_directors.csv", sep = ";")



actors.head(2)

directors.head(2)

# Asking for the password of the MySQL database.

password = getpass("MySQL password: ")


# Creating a variable called db_name and assigning it the value of "Netflix"
db_name = "Netflix"

# Creating a connection string to the database.
conexion = f"mysql+pymysql://root:{password}@localhost/{db_name}"

# Creating an engine to connect to the database.
engine = alch.create_engine(conexion)


# Insert data

## Actors

actors.head()

# Replacing all the single quotes with nothing.
actors["description"] = actors["description"].str.replace("'", "")
actors["name"] = actors["name"].str.strip()

for index, row in actors.iterrows():
    
    # Selecting the idActors from the actors table where the name is equal to the name of the row.
    actors_id = engine.execute(f"""select idActors from actors WHERE name = "{row['name']}";""")

    if len(list(actors_id)) > 0: 
        print("sorry this id is already in our database")
        
    else: 
        engine.execute(f"""
            INSERT INTO actors (name, description, positions) VALUES 
            ( "{row['name']}", '{row['description']}', "{row['positions']}")
            """) 


## Directors

directors.head(2)

directors["description"] = directors["description"].str.replace("'", "")
directors["name"] = directors["name"].str.strip()

for index, row in directors.iterrows():
    
    
    # Selecting the idActors from the actors table where the name is equal to the name of the row.
    directors_id = engine.execute(f"""select iddirectors from directors WHERE director = "{row['name']}";""")

    
    
    
    if len(list(directors_id)) > 0: 
        print("sorry this id is already in our database")
        
    else: 
        print(f"""
            INSERT INTO directors (director, description, positions) VALUES 
            ( "{row['name']}", '{row['description']}', "{row['positions']}")
            """)
        engine.execute(f"""
            INSERT INTO directors (director, description, positions) VALUES 
            ( "{row['name']}", '{row['description']}', "{row['positions']}")
            """)
